+++ lib/util/leaveGroup.js
@@ -1,5 +1,5 @@
 // Includes
-var generalRequest = require('./util/generalRequest.js').func;
+var generalRequest = require('./generalRequest.js').func;
 
 // Args
 exports.required = ['group'];
